Mac controls, Tutorial, Balancing
Spent some time looking into places and ways to share my game to get some user input.  Ended up having one friend who’s whole family only has Mac, so I started implementing new controls.
- Struggled with adding a second control to the scroll effect I created.  Managed to find a fairly solid solution with a coroutine to slow the UI selection changes from mouse movement.
- Ended up learning some things about the input system, and also Mac computers.  It turns out I know dangerously little about them, so I think I’ll buy an older one right away to play with and test on.
- Added scripts to switch some controls text for Mac users
- Balanced the Damage and Knockback Upgrades as they were OP
- Slowed the baseline for subwave spawn time
- Created basic tutorial popups for a welcome screen, build system, and upgrade system.
- Changed UIUpgrades to update their text in coroutine every 1/5 second.
- Due to the new Standard WASD movement type, adjusted movement calculation to smooth movement speed and also add a buffer to the angle toward the mouse which full speed is achieved.

+++ Mutant Mayhem/Assets/_Scripts/Weapons/GunSO.cs
@@ -30,8 +30,8 @@ public class GunSO : ScriptableObject
     public float muzzleFlashTime;
     
     [Header("Gun Stats")]
-    public int damage;
-    public int damageUpgAmt;
+    public float damage;
+    public float damageUpgAmt;
     public float knockback;
     public float knockbackUpgAmt;
     public int clipSize;
